apps/task/serializers.py
from rest_framework import serializers
from .models import Task
from django.contrib.auth.models import User

class TaskSerializer(serializers.ModelSerializer):
    """Сериализатор для задачи"""
    is_overdue = serializers.ReadOnlyField()

    class Meta:
        model = Task
        fields = '__all__'
        read_only_fields = ('user',)


class RegisterSerializer(serializers.ModelSerializer):
    """Сериализатор для регистрации пользователя"""
    class Meta:
        model = User
        fields = ['username', 'email', 'password']
        extra_kwargs = {'password': {'write_only': True}}

    def create(self, validated_data):
        return User.objects.create_user(**validated_data)